+++ pkg/hooks/docker/hook_test.go
@@ -1,11 +1,7 @@
 package docker
 
 import (
-	"bytes"
-	"errors"
 	"io/ioutil"
-	"net/http"
-	"net/http/httptest"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
@@ -55,28 +51,6 @@ func TestParse(t *testing.T) {
 	}
 }
 
-func TestParseWithNoBody(t *testing.T) {
-	bodyErr := errors.New("just a test error")
-
-	req := httptest.NewRequest("POST", "/", failingReader{err: bodyErr})
-
-	_, err := Parse(req)
-	if err != bodyErr {
-		t.Fatal("expected an error")
-	}
-
-}
-
-func TestParseWithUnparseableBody(t *testing.T) {
-	req := httptest.NewRequest("POST", "/", nil)
-
-	_, err := Parse(req)
-
-	if err == nil {
-		t.Fatal("expected an error")
-	}
-}
-
 func TestPushedImageURL(t *testing.T) {
 	hook := &Webhook{
 		PushData: &PushData{
@@ -109,24 +83,11 @@ func TestRepository(t *testing.T) {
 	}
 }
 
-func makeHookRequest(t *testing.T, fixture string) *http.Request {
+func makeHookRequest(t *testing.T, fixture string) []byte {
 	t.Helper()
 	b, err := ioutil.ReadFile(fixture)
 	if err != nil {
 		t.Fatalf("failed to read %s: %s", fixture, err)
 	}
-	req := httptest.NewRequest("POST", "/", bytes.NewReader(b))
-	req.Header.Add("Content-Type", "application/json")
-	return req
-}
-
-type failingReader struct {
-	err error
-}
-
-func (f failingReader) Read(p []byte) (n int, err error) {
-	return 0, f.err
-}
-func (f failingReader) Close() error {
-	return f.err
+	return b
 }
