+++ pkg/hooks/gcr/hook_test.go
@@ -1,11 +1,7 @@
 package gcr
 
 import (
-	"bytes"
-	"errors"
 	"io/ioutil"
-	"net/http"
-	"net/http/httptest"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
@@ -31,28 +27,6 @@ func TestParse(t *testing.T) {
 	}
 	if diff := cmp.Diff(want, hook); diff != "" {
 		t.Fatalf("hook doesn't match:\n%s", diff)
-	}
-}
-
-func TestParseWithNoBody(t *testing.T) {
-	bodyErr := errors.New("just a test error")
-
-	req := httptest.NewRequest("POST", "/", failingReader{err: bodyErr})
-
-	_, err := Parse(req)
-	if err != bodyErr {
-		t.Fatal("expected an error")
-	}
-
-}
-
-func TestParseWithUnparseableBody(t *testing.T) {
-	req := httptest.NewRequest("POST", "/", nil)
-
-	_, err := Parse(req)
-
-	if err == nil {
-		t.Fatal("expected an error")
 	}
 }
 
@@ -82,15 +56,13 @@ func TestRepository(t *testing.T) {
 	}
 }
 
-func makeHookRequest(t *testing.T, fixture string) *http.Request {
+func makeHookRequest(t *testing.T, fixture string) []byte {
 	t.Helper()
 	b, err := ioutil.ReadFile(fixture)
 	if err != nil {
 		t.Fatalf("failed to read %s: %s", fixture, err)
 	}
-	req := httptest.NewRequest("POST", "/", bytes.NewReader(b))
-	req.Header.Add("Content-Type", "application/json")
-	return req
+	return b
 }
 
 type failingReader struct {
