+++ pkg/hooks/interface.go
@@ -1,7 +1,5 @@
 package hooks
 
-import "net/http"
-
 // PushEvent values return the image that is to be inserted into the file.
 type PushEvent interface {
 	PushedImageURL() string
@@ -10,4 +8,4 @@ type PushEvent interface {
 }
 
 // PushEventParser parses the specifics of a hook request into a body.
-type PushEventParser func(req *http.Request) (PushEvent, error)
+type PushEventParser func(payload []byte) (PushEvent, error)
